+++ src/test/java/ladder/domain/ResultTest.java
@@ -1,9 +1,8 @@
-package ladder.domain.strategy.linestrategy;
+package ladder.domain;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 
-import ladder.domain.Result;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
