+++ src/test/java/ladder/domain/ResultsTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 
-import ladder.domain.linestrategy.CustomLineStrategy;
+import ladder.domain.strategy.linestrategy.CustomLineStrategy;
 import ladder.domain.strategy.linestrategy.LineStrategy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
